Open Olympiad Arena's doors
JoeAlisson/L2jOrg#93

+++ Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/network/serverpackets/OnEventTrigger.java
@@ -38,7 +38,7 @@ public class OnEventTrigger extends ServerPacket {
         writeId(ServerPacketId.EVENT_TRIGGER);
 
         writeInt(_emitterId);
-        writeByte((byte) _enabled);
+        writeByte(_enabled);
     }
 
 }
